requirements.txt
curl-cffi
google-genai
python-dotenv
pandas